proxy: use HTTP GET for /health handler
Currently we use a v3 Get to check for endpoint /health. The downside of this is that server metrics do not properly record the health check but instead a gRPC Get. This change would alter the handler from doing a clientv3 Get to a HTTP GET against etcd server endpoint /health.
