+++ proxy/grpcproxy/metrics.go
@@ -17,10 +17,7 @@ package grpcproxy
 import (
 	"fmt"
 	"io/ioutil"
-	"math/rand"
 	"net/http"
-	"strings"
-	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
 	"go.etcd.io/etcd/etcdserver/api/etcdhttp"
@@ -69,23 +66,9 @@ func init() {
 
 // HandleMetrics performs a GET request against etcd endpoint and returns '/metrics'.
 func HandleMetrics(mux *http.ServeMux, c *http.Client, eps []string) {
-	// random shuffle endpoints
-	r := rand.New(rand.NewSource(int64(time.Now().Nanosecond())))
-	if len(eps) > 1 {
-		eps = shuffleEndpoints(r, eps)
-	}
-
 	pathMetrics := etcdhttp.PathMetrics
 	mux.HandleFunc(pathMetrics, func(w http.ResponseWriter, r *http.Request) {
-		target := fmt.Sprintf("%s%s", eps[0], pathMetrics)
-		if !strings.HasPrefix(target, "http") {
-			scheme := "http"
-			if r.TLS != nil {
-				scheme = "https"
-			}
-			target = fmt.Sprintf("%s://%s", scheme, target)
-		}
-
+		target := createTarget(eps, pathMetrics, r.TLS, true)
 		resp, err := c.Get(target)
 		if err != nil {
 			http.Error(w, "Internal server error", http.StatusInternalServerError)
@@ -95,20 +78,4 @@ func HandleMetrics(mux *http.ServeMux, c *http.Client, eps []string) {
 		body, _ := ioutil.ReadAll(resp.Body)
 		fmt.Fprintf(w, "%s", body)
 	})
-}
-
-func shuffleEndpoints(r *rand.Rand, eps []string) []string {
-	// copied from Go 1.9<= rand.Rand.Perm
-	n := len(eps)
-	p := make([]int, n)
-	for i := 0; i < n; i++ {
-		j := r.Intn(i + 1)
-		p[i] = p[j]
-		p[j] = i
-	}
-	neps := make([]string, n)
-	for i, k := range p {
-		neps[i] = eps[k]
-	}
-	return neps
 }
